refactor(UI): move load models in separate folder
