+++ ScriptBeeClient/src/app/pages/projects/project-details/model/load-model-page/link-models/link-models.component.ts
@@ -1,20 +1,23 @@
 import { Component, computed, inject, input, signal } from '@angular/core';
-import { CenteredSpinnerComponent } from '../../../../../components/centered-spinner/centered-spinner.component';
-import { ErrorStateComponent } from '../../../../../components/error-state/error-state.component';
+import { ErrorStateComponent } from '../../../../../../components/error-state/error-state.component';
 import { MatFormField, MatLabel } from '@angular/material/form-field';
 import { MatOption } from '@angular/material/core';
 import { MatSelect } from '@angular/material/select';
 import { rxResource } from '@angular/core/rxjs-interop';
-import { LinkerService } from '../../../../../services/linkers/linker.service';
+import { LinkerService } from '../../../../../../services/linkers/linker.service';
 import { MatButton } from '@angular/material/button';
 import { finalize } from 'rxjs';
-import { convertError } from '../../../../../utils/api';
+import { convertError } from '../../../../../../utils/api';
+import { LoadingProgressBarComponent } from '../../../../../../components/loading-progress-bar/loading-progress-bar.component';
+import { HttpErrorResponse } from '@angular/common/http';
+import { RouterStateSnapshot } from '@angular/router';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-link-models',
   templateUrl: './link-models.component.html',
   styleUrls: ['./link-models.component.scss'],
-  imports: [CenteredSpinnerComponent, ErrorStateComponent, MatFormField, MatLabel, MatOption, MatSelect, MatButton],
+  imports: [ErrorStateComponent, MatFormField, MatLabel, MatOption, MatSelect, MatButton, LoadingProgressBarComponent],
 })
 export class LinkModelsComponent {
   projectId = input.required<string>();
@@ -34,6 +37,7 @@ export class LinkModelsComponent {
   isLinkLoading = signal(false);
 
   private linkerService = inject(LinkerService);
+  private snackbar = inject(MatSnackBar);
 
   onLinkButtonClick() {
     const linkerId = this.selectedLinkerId();
@@ -45,6 +49,14 @@ export class LinkModelsComponent {
     this.linkerService
       .linkModels(this.projectId(), this.instanceId(), linkerId)
       .pipe(finalize(() => this.isLinkLoading.set(false)))
-      .subscribe();
+      .subscribe({
+        next: () => {
+          this.snackbar.open('Link successful', 'Dismiss', { duration: 4000 });
+        },
+        error: (errorResponse: HttpErrorResponse) => {
+          const error = convertError(errorResponse);
+          this.snackbar.open(`Could not link models because ${error?.title}`, 'Dismiss', { duration: 4000 });
+        },
+      });
   }
 }
