+++ ScriptBeeClient/src/app/pages/projects/project-details/model/load-model-page/load-models/load-models.component.scss
@@ -1,0 +1,5 @@
+.load-button {
+  margin-top: 16px;
+  margin-bottom: 24px;
+  margin-left: 8px;
+}
